Add strategy asset optimization, symbol selection, crypto support, and Coinbase fixes
- Categorize strategies by asset optimization (stocks/crypto/generic) and filter
  by broker compatibility in API and bot creation form
- Add trading pairs service with curated Coinbase crypto and Alpaca stock lists
- Add symbols column to bot_configurations (migration 008) for user-selected pairs
- Pass symbols through Redis start command to engine
- Add sigma_alpha_bull_crypto strategy optimized for crypto markets
- Reorder bot creation: pick broker first, then see compatible strategies
- Add search and optimization filter to strategies page
- Fix Coinbase SDK response parsing (typed objects, not dicts) in data layer
- Fix Coinbase broker order side casing (uppercase BUY/SELL)
- Handle failed CreateOrderResponse (extract error vs success response)
- Add Coinbase order smoke test for validating full order lifecycle
- Update strategy configs: disable stock strategies, enable crypto for Coinbase

+++ services/trading-engine/trading_automata/strategies/examples/rsi_atr_trend.py
@@ -99,7 +99,7 @@ class RSIATRTrendStrategy(BaseStrategy):
         self.price_history[bar.symbol].append(float(bar.close))
         self.high_history[bar.symbol].append(float(bar.high))
         self.low_history[bar.symbol].append(float(bar.low))
-        self.volume_history[bar.symbol].append(int(bar.volume))
+        self.volume_history[bar.symbol].append(float(bar.volume))
 
         # Need enough data for indicators
         min_required = max(self.rsi_period, self.ema_slow_period, self.atr_period) + 5
